@@ -39,24 +39,23 @@ class Settings extends Controller {
 		$this->render( 'admin/settings.html' );
 	}
 
-
 	protected function handlePost() {
 		$this->form->requireInt( 'phase1pass' );
 		$this->form->requireInt( 'phase2pass' );
 		$this->form->requireFloat( 'relexp' );
 		$this->form->requireFloat( 'expshare' );
 		$this->form->requireDateTime( 'apply_open', 'Y-m-d' );
-		$this->form->requireDateTime( 'apply_close' , 'Y-m-d' );
+		$this->form->requireDateTime( 'apply_close', 'Y-m-d' );
 
 		if ( $this->form->validate() ) {
-			$settings = array(
+			$settings = [
 				'phase1pass' => $this->form->get( 'phase1pass' ),
 				'phase2pass' => $this->form->get( 'phase2pass' ),
 				'relexp' => $this->form->get( 'relexp' ),
 				'expshare' => $this->form->get( 'expshare' ),
 				'apply_open' => $this->form->get( 'apply_open' )->format( 'Y-m-d' ),
 				'apply_close' => $this->form->get( 'apply_close' )->format( 'Y-m-d' ),
-			);
+			];
 
 			if ( ( $settings['relexp'] + $settings['expshare'] ) != 1 ) {
 				$this->flash( 'error', 'Sum of weights must be one' );
@@ -71,7 +70,7 @@ class Settings extends Controller {
 				}
 			}
 		} else {
-			//FIXME: actually pass form errors back to view
+			// FIXME: actually pass form errors back to view
 			$this->flash( 'error', 'Invalid input.' );
 		}
 
