@@ -32,12 +32,12 @@ if ( !defined( 'APP_ROOT' ) ) {
 require_once APP_ROOT . '/vendor/autoload.php';
 
 // Ensure that a default timezone is set
-set_error_handler(function ($errno, $errstr) {
+set_error_handler( function ( $errno, $errstr ) {
 	throw new Exception( $errstr );
 } );
 try {
 	date_default_timezone_get();
-} catch( Exception $e ) {
+} catch ( Exception $e ) {
 	// Use UTC if not specified anywhere in .ini
 	date_default_timezone_set( 'UTC' );
 }
